Strictly use 32 bit integers in optical flow least squares
This patch remove int64_t in optical flow least squares solvers. Two changes are included:

1. Replace int64_t by int in the matrix filling stage in OPFL and DAMR. This change has no stats difference.
2. Fixes in optical flow bit range checks, which prevents the intermediate results from getting more than 32 bits.

   This change has noise level BD rate difference (\<0.01) in CTC.

STATS_CHANGED

+++ test/opt_flow_test.cc
@@ -869,7 +869,7 @@ typedef void (*calc_affine_autocorrelation_matrix)(
 #if CONFIG_AFFINE_REFINEMENT_SB
     int x_offset, int y_offset,
 #endif  // CONFIG_AFFINE_REFINEMENT_SB
-    int64_t *mat_a, int64_t *vec_b);
+    int32_t *mat_a, int32_t *vec_b);
 
 class AV1CalcAffineAutocorrelationMatrixTest
     : public AV1OptFlowTest<calc_affine_autocorrelation_matrix> {
@@ -942,10 +942,10 @@ class AV1CalcAffineAutocorrelationMatrixTest
                                    const int16_t *pdiff, int pstride,
                                    const int16_t *gx, const int16_t *gy,
                                    int gstride, int bw, int bh) {
-    DECLARE_ALIGNED(32, int64_t, mat_ref[16]);
-    DECLARE_ALIGNED(32, int64_t, mat_test[16]);
-    DECLARE_ALIGNED(32, int64_t, vec_ref[4]);
-    DECLARE_ALIGNED(32, int64_t, vec_test[4]);
+    DECLARE_ALIGNED(32, int32_t, mat_ref[16]);
+    DECLARE_ALIGNED(32, int32_t, mat_test[16]);
+    DECLARE_ALIGNED(32, int32_t, vec_ref[4]);
+    DECLARE_ALIGNED(32, int32_t, vec_test[4]);
     memset(mat_ref, 0, sizeof(mat_ref));
     memset(mat_test, 0, sizeof(mat_test));
     memset(vec_ref, 0, sizeof(vec_ref));
@@ -965,8 +965,7 @@ class AV1CalcAffineAutocorrelationMatrixTest
     for (int i = 0; i < 16; ++i) {
       if (mat_ref[i] != mat_test[i]) {
         failed = 1;
-        printf("Mat [%4d] ref %6" PRId64 " test %6" PRId64 " \n", i, mat_ref[i],
-               mat_test[i]);
+        printf("Mat [%4d] ref %d test %d \n", i, mat_ref[i], mat_test[i]);
         break;
       }
     }
@@ -974,8 +973,7 @@ class AV1CalcAffineAutocorrelationMatrixTest
     for (int i = 0; i < 4; ++i) {
       if (vec_ref[i] != vec_test[i]) {
         failed = 1;
-        printf("Vec [%4d] ref %6" PRId64 " test %6" PRId64 " \n", i, vec_ref[i],
-               vec_test[i]);
+        printf("Vec [%4d] ref %d test %d \n", i, vec_ref[i], vec_test[i]);
         break;
       }
     }
@@ -987,10 +985,10 @@ class AV1CalcAffineAutocorrelationMatrixTest
       calc_affine_autocorrelation_matrix test_func, const int16_t *pdiff,
       int pstride, const int16_t *gx, const int16_t *gy, int gstride, int bw,
       int bh) {
-    DECLARE_ALIGNED(32, int64_t, mat_ref[16]);
-    DECLARE_ALIGNED(32, int64_t, mat_test[16]);
-    DECLARE_ALIGNED(32, int64_t, vec_ref[4]);
-    DECLARE_ALIGNED(32, int64_t, vec_test[4]);
+    DECLARE_ALIGNED(32, int32_t, mat_ref[16]);
+    DECLARE_ALIGNED(32, int32_t, mat_test[16]);
+    DECLARE_ALIGNED(32, int32_t, vec_ref[4]);
+    DECLARE_ALIGNED(32, int32_t, vec_test[4]);
     memset(mat_ref, 0, sizeof(mat_ref));
     memset(mat_test, 0, sizeof(mat_test));
     memset(vec_ref, 0, sizeof(vec_ref));
